requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.22.0

# Azure AI services
azure-ai-document-intelligence>=1.0.0
azure-storage-blob>=12.16.0
azure-cosmos>=4.5.0
azure-identity>=1.13.0
azure-ai-openai>=1.0.0

# Semantic Kernel
semantic-kernel>=0.3.0.dev
semantic-kernel-connectors-ai-openai>=0.3.0.dev

# AutoGen framework
pyautogen>=0.2.0
autogenstudio>=0.0.39

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0
openpyxl>=3.1.2

# Document processing
pypdf2>=3.0.0
python-docx>=0.8.11
azure-ai-formrecognizer>=3.2.0

# Testing tools
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Logging and monitoring
structlog>=23.1.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-otlp>=1.18.0

# Development tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0